*: Add "static" and "const" keywords and fix type mismatch
Noteworthy changes are as follows:

- sent_size in struct sock_buf is changed from int to size_t.

- Comment out the following unused functions: get_bhyve_env(),
  next_bhyve_env(), and get_bhyve_env_env().

- Use strdup() for key in struct conf_var.  It was set in set_var{,0}()
  and the key was often initialized using the pointer of a string literal.
  This is a bug because free_var() calls free(3) for the pointer.
  Also, the type of key is incompatible with a string literal, whose
  type is "const char *".

- Expose set_string() to set grub_run_partition in struct inspection.
  Currently inspect_netbsd_iso() uses it.  free_inspection() now calls
  free(3) for consistency.

- md_unit in struct inspection is changed from unsigned int to int.

- Eliminate assingments to argv[1] in control.c.  It is incompatible with
  the type of string literals.

- Use static memory for fields of gl_conf0, not read-only memory.

+++ bmd.h
@@ -122,7 +122,7 @@ struct sock_buf {
 	size_t read_size;
 	size_t read_bytes;
 	char *buf;
-	int sent_size;
+	size_t sent_size;
 	int res_fd;
 	size_t res_size;
 	size_t res_bytes;
